requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.10.0=py37hc8dfbb8_1
argon2-cffi=20.1.0=py37h4ab8f01_2
astor=0.8.1=pyh9f0ad1d_0
async_generator=1.10=py_0
attrs=20.2.0=pyh9f0ad1d_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.1=py_0
bleach=3.2.1=pyh9f0ad1d_0
brotlipy=0.7.0=py37h0013d47_1001
ca-certificates=2020.6.20=hecda079_0
certifi=2020.6.20=py37hf50a25e_2
cffi=1.14.3=py37hd6b71e5_1
chardet=3.0.4=py37hf50a25e_1008
colorama=0.4.4=pyh9f0ad1d_0
cryptography=3.2.1=py37hd8e9650_0
cycler=0.10.0=py_2
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
entrypoints=0.3=py37hc8dfbb8_1002
freeglut=3.0.0=h6538335_1005
freetype=2.10.4=h546665d_0
gast=0.4.0=pyh9f0ad1d_0
google-pasta=0.2.0=pyh8c360ce_0
grpcio=1.33.2=py37h04d2302_2
h5py=2.10.0=nompi_py37h23cfb99_105
hdf5=1.10.6=nompi_h89124ea_1110
hurry.filesize=0.9=pyh8c360ce_0
icu=67.1=h33f27b4_0
idna=2.10=pyh9f0ad1d_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
intel-openmp=2020.2=254
ipykernel=5.3.4=py37h7b7c402_1
ipython=7.19.0=py37heaed05f_0
ipython_genutils=0.2.0=py_1
jasper=2.0.14=hdc05fd1_1
jedi=0.17.2=py37hc8dfbb8_1
jinja2=2.11.2=pyh9f0ad1d_0
joblib=0.17.0=py_0
jpeg=9d=h8ffe710_0
jsonschema=3.2.0=py_2
jupyter_client=6.1.7=py_0
jupyter_core=4.6.3=py37hc8dfbb8_2
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
keras=2.3.1=py37h21ff451_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.0=py_0
kiwisolver=1.3.1=py37h8c56517_0
krb5=1.17.1=hc04afaa_3
libblas=3.8.0=20_mkl
libcblas=3.8.0=20_mkl
libclang=10.0.1=default_hf44288c_1
libcurl=7.71.1=h4b64cdc_8
libgpuarray=0.7.6=hfa6e2cd_1003
liblapack=3.8.0=20_mkl
liblapacke=3.8.0=20_mkl
libopencv=4.5.0=py37_2
libpng=1.6.37=h1d00b33_2
libprotobuf=3.13.0.1=h200bbdf_0
libsodium=1.0.18=h62dcd97_1
libssh2=1.9.0=hb06d900_5
libtiff=4.1.0=hc10be44_6
libwebp-base=1.1.0=h8ffe710_3
lz4-c=1.9.2=h62dcd97_2
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
mako=1.1.3=pyh9f0ad1d_0
markdown=3.3.3=pyh9f0ad1d_0
markupsafe=1.1.1=py37h0013d47_2
matplotlib=3.3.2=py37hc8dfbb8_1
matplotlib-base=3.3.2=py37h64d6e1c_1
mistune=0.8.4=py37h4ab8f01_1002
mkl=2020.2=256
msys2-conda-epoch=20160418=1
nbclient=0.5.1=py_0
nbconvert=6.0.7=py37hc8dfbb8_2
nbformat=5.0.8=py_0
nest-asyncio=1.4.2=pyhd8ed1ab_0
notebook=6.1.4=py37hc8dfbb8_1
numpy=1.19.2=py37he804e03_1
olefile=0.46=pyh9f0ad1d_1
opencv=4.5.0=py37_2
openssl=1.1.1h=he774522_0
packaging=20.4=pyh9f0ad1d_0
pandas=1.1.4=py37h08fd248_0
pandoc=2.11.0.4=hf4a77e7_0
pandocfilters=1.4.2=py_1
parso=0.7.1=pyh9f0ad1d_0
pickleshare=0.7.5=py_1003
pillow=8.0.1=py37hb71ed58_0
pip=20.2.4=py_0
prometheus_client=0.8.0=pyh9f0ad1d_0
prompt-toolkit=3.0.8=py_0
protobuf=3.13.0.1=py37h1834ac0_1
py-opencv=4.5.0=py37hc6149b9_2
pycparser=2.20=pyh9f0ad1d_2
pygments=2.7.2=py_0
pygpu=0.7.6=py37h3a3b6f7_1002
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.12.3=py37h1834ac0_4
pyqt5-sip=4.19.18=pypi_0
pyqtchart=5.12=pypi_0
pyqtwebengine=5.12.1=pypi_0
pyreadline=2.1=py37hc8dfbb8_1002
pyrsistent=0.17.3=py37h4ab8f01_1
pysocks=1.7.1=py37hf50a25e_2
python=3.7.8=h60c2a47_1_cpython
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytz=2020.4=pyhd8ed1ab_0
pywin32=228=py37h4ab8f01_0
pywinpty=0.5.7=py37hc8dfbb8_1
pyyaml=5.3.1=py37h0013d47_1
pyzmq=19.0.2=py37h453f00a_2
qt=5.12.9=hb2cf2c5_0
requests=2.24.0=pyh9f0ad1d_0
scikit-learn=0.23.2=py37heb15398_2
scipy=1.5.3=py37h6db1a17_0
send2trash=1.5.0=py_0
setuptools=49.6.0=py37hf50a25e_2
six=1.15.0=pyh9f0ad1d_0
sqlite=3.33.0=he774522_1
tensorboard=1.14.0=py37_0
tensorflow=1.14.0=h1f41ff6_0
tensorflow-base=1.14.0=py37hc8dfbb8_0
tensorflow-estimator=1.14.0=py37h5ca1d4c_0
termcolor=1.1.0=py_2
terminado=0.9.1=py37hc8dfbb8_1
testpath=0.4.4=py_0
theano=1.0.5=py37hf2a7229_1
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=he774522_1
tornado=6.1=py37hcc03f2d_0
traitlets=5.0.5=py_0
urllib3=1.25.11=py_0
vc=14.1=h869be7e_1
vs2015_runtime=14.16.27012=h30e32a0_2
vs2017_win-64=19.16.27038=h2e3bad8_2
vswhere=2.7.1=h21ff451_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
werkzeug=1.0.1=pyh9f0ad1d_0
wheel=0.35.1=pyh9f0ad1d_0
win_inet_pton=1.1.0=py37hc8dfbb8_1
wincertstore=0.2=py37hc8dfbb8_1005
winpty=0.4.3=4
wrapt=1.12.1=py37h8055547_1
xz=5.2.5=h62dcd97_1
yaml=0.2.5=he774522_0
zeromq=4.3.2=ha925a31_4
zipp=3.4.0=py_0
zlib=1.2.11=h62dcd97_1010
zstd=1.4.5=h1f3a1b7_2
